Add comprehensive validation layer with FluentValidation, update dependent use cases, refactor MongoDB adapters for better pagination handling, and improve validation robustness in vehicle and rental use cases.

+++ src/GtMotive.Estimate.Microservice.ApplicationCore/Common/PagedResponse.cs
@@ -10,7 +10,7 @@
 /// <param name="PageSize">The size of the page.</param>
 public record PagedResponse<T>(
     List<T> Data,
-    int TotalCount,
+    long TotalCount,
     int PageNumber,
     int PageSize
 )
