+++ src/GtMotive.Estimate.Microservice.Domain/Interfaces/Port/IRentVehiclePort.cs
@@ -7,7 +7,7 @@ public interface IRentVehiclePort : IUnitOfWork
 {
     Task<RentInformation?> GetVehicleRent(Expression<Func<RentInformation, bool>> filter);
 
-    Task<(List<RentInformation>, int)> GetVehiclesRent(Expression<Func<RentInformation, bool>> filter, int pageIndex,
+    Task<(List<RentInformation>, long)> GetVehiclesRent(Expression<Func<RentInformation, bool>> filter, int pageIndex,
         int pageSize);
 
     Task<List<RentInformation>> GetVehiclesRent(Expression<Func<RentInformation, bool>> filter);
